requirements.txt
asgiref==3.5.0
bson==0.5.10
certifi==2021.10.8
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cors==1.0.1
distlib==0.3.4
Django==4.0.2
dnspython==2.2.1
filelock==3.4.2
Flask==2.2.2
Flask-Cors==3.0.10
future==0.18.2
geographiclib==2.0
geopy==2.3.0
gevent==22.10.2
googlemaps==4.7.3
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.5.0
pipenv==2022.1.8
platformdirs==2.4.1
pymongo==4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.1
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.2
tldextract==3.4.0
urllib3==1.26.13
virtualenv==20.13.0
virtualenv-clone==0.5.7
Werkzeug==2.2.2
zope.event==4.6
zope.interface==5.5.2
